LayoutTests/fast/js/sputnik/Conformance/11_Expressions/11.1_Primary_Expressions/11.1.5_Object_Initializer/S11.1.5_A4.1-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Parse error
S11.1.5_A4.1

PASS Expected parsing failure

TEST COMPLETE
